// lib/clientSideScripts/customCss.interfaces.ts
export interface CssOptions {
  // The address bar padding that needs to be added
  addressBarPadding: number;
  // En/disable the animations
  disableCSSAnimation: boolean;
  // The id that can identify the css style
  id: string;
  // The tool bar padding that needs to be added
  toolBarPadding: number;
}